Test update

+++ vue/src/views/Events.vue
@@ -2,12 +2,13 @@
   <div style="padding: 10px">
     <!--    功能区-->
     <div style="margin: 10px 0">
-      <el-button type="primary" @click="add">
+      <el-button type="primary" @click="add" v-if="user.role ==='ROLE_ADMIN'|| user.role ==='ROLE_USER'">
         <el-icon>
           <CirclePlus/>
         </el-icon>
         &nbsp发表新事件
       </el-button>
+
     </div>
     <!--    搜索区-->
     <div style="margin: 10px 0;display: block;clear: both">
@@ -15,44 +16,49 @@
                 clearable/>
       <el-input v-model="author" placeholder="请输入作者" style="width: 20%" :prefix-icon="Search" class="mr-10"
                 clearable/>
-      <!--      <el-input v-model="createTime" placeholder="请选择时间" style="width: 20%" :prefix-icon="Search" class="mr-10"-->
-      <!--                clearable/>-->
-      <!--      <el-date-picker-->
-      <!--          class="mr-10"-->
-      <!--          v-model="createTime"-->
-      <!--          type="daterange"-->
-      <!--          unlink-panels-->
-      <!--          range-separator="-"-->
-      <!--          start-placeholder="开始日期"-->
-      <!--          end-placeholder="结束日期"-->
-      <!--          :shortcuts="shortcuts"-->
-      <!--          :size="size"-->
-      <!--      />-->
+      <el-date-picker
+          class="mr-10"
+          v-model="timeHorizon"
+          type="daterange"
+          range-separator="至"
+          start-placeholder="开始日期"
+          end-placeholder="结束日期"
+          unlink-panels
+      />
       <el-button type="primary" class="mb-10" @click="load">查询</el-button>
       <el-button type="primary" class="mb-10" @click="reset">重置</el-button>
-      <el-button type="danger" style="float: right;margin-right: 10px">
-        <el-icon>
-          <Delete/>
-        </el-icon>
-        &nbsp批量删除
-      </el-button>
+      <el-popconfirm title="确定要删除吗" @confirm="deleteBatch">
+        <template #reference>
+          <el-button class="mb-10" type="danger" style="float: right;margin-right: 10px"
+                     v-if="user.role ==='ROLE_ADMIN'">
+            <el-icon>
+              <Delete/>
+            </el-icon>
+            &nbsp批量删除
+          </el-button>
+        </template>
+      </el-popconfirm>
+
 
     </div>
     <!--    列表区-->
     <!--        stripe:斑马纹-->
     <el-table :data="tableData" border stripe style="width: 100%" @selection-change="handleSelectionChange"
               highlight-current-row>
-      <el-table-column type="selection" width="40px"/>
+      <el-table-column type="selection" width="40px" align="center"/>
       <!--            sortable:排序操作-->
-      <el-table-column prop="id" label="ID" sortable=""/>
-      <el-table-column prop="title" label="标题"/>
-      <el-table-column prop="author" label="作者"/>
-      <el-table-column prop="createTime" label="创建时间"/>
-      <el-table-column fixed="right" label="操作" width="300px">
+      <el-table-column prop="id" label="ID" sortable="" align="center" width="70px"/>
+      <el-table-column prop="title" label="标题" align="center"/>
+      <el-table-column prop="author" label="作者" align="center"/>
+      <el-table-column prop="createTime" label="创建时间" align="center"/>
+      <el-table-column fixed="right" label="操作" width="300px" align="center">
         <template #default="scope">
           <el-button plain type="success" @click="details(scope.row)">详情</el-button>
-          <el-button plain type="primary" @click="handleEdit(scope.row)">编辑</el-button>
-          <el-popconfirm title="你确定要删除吗?" @confirm="handleDelete(scope.row)">
+          <el-button plain type="primary" @click="handleEdit(scope.row)"
+                     v-if="scope.row.author===this.user.name">编辑
+          </el-button>
+          <el-popconfirm title="你确定要删除吗?" @confirm="handleDelete(scope.row)"
+                         v-if="user.role ==='ROLE_ADMIN'|| user.role ==='ROLE_USER'">
             <template #reference>
               <el-button type="danger">
                 <el-icon>
@@ -79,10 +85,11 @@
           @current-change="handleCurrentChange"
       />
       <el-dialog v-model="dialogVisible" title="编辑事件" width="50%">
-        <el-form model="form" label-width="120px">
-          <el-form-item label="标题">
+        <el-form :model="form" label-width="120px" :rules="rules" ref="pass">
+          <el-form-item label="标题" prop="title">
             <el-input v-model="form.title" style="width: 50%"/>
           </el-form-item>
+          <!--                    正文-->
           <div id="div1">
 
           </div>
@@ -90,7 +97,7 @@
         <template #footer>
                     <span class="dialog-footer">
                         <el-button @click="dialogVisible = false">取消</el-button>
-                        <el-button type="primary" @click="save">发布</el-button>
+                        <el-button type="primary" @click="save" :loading="loading">发布</el-button>
                     </span>
         </template>
       </el-dialog>
@@ -111,56 +118,35 @@ import {Search, Delete} from "@element-plus/icons-vue";
 import '@wangeditor/editor/dist/css/style.css' // 引入 css
 
 import E from 'wangeditor'
+import {serverIp} from "../../public/config";
 //设置全局变量
 let editor;
-let userStr = sessionStorage.getItem("user") || "{}"
-let user = JSON.parse(userStr)
-// const shortcuts = [
-//   {
-//     text: 'Last week',
-//     value: () => {
-//       const end = new Date()
-//       const start = new Date()
-//       start.setTime(start.getTime() - 3600 * 1000 * 24 * 7)
-//       return [start, end]
-//     },
-//   },
-//   {
-//     text: 'Last month',
-//     value: () => {
-//       const end = new Date()
-//       const start = new Date()
-//       start.setTime(start.getTime() - 3600 * 1000 * 24 * 30)
-//       return [start, end]
-//     },
-//   },
-//   {
-//     text: 'Last 3 months',
-//     value: () => {
-//       const end = new Date()
-//       const start = new Date()
-//       start.setTime(start.getTime() - 3600 * 1000 * 24 * 90)
-//       return [start, end]
-//     },
-//   },
-// ]
+
 export default {
   name: 'Events',
-  components: {},
   data() {
     return {
+      loading: false,
       form: {},
       dialogVisible: false,
       title: '',
       author: '',
-      // createTime: '',
+      timeHorizon: [],
       currentPage: 1,
       pageSize: 10,
       total: 10,
       tableData: [],
       detail: {},
       vis: false,
-      editor: null
+      editor: null,
+      ids: [],
+      user: sessionStorage.getItem("user") ? JSON.parse(sessionStorage.getItem("user")) : {},
+      rules: {
+        title: [
+          {required: true, message: '标题不能为空', trigger: 'blur'},
+          {min: 1, max: 10, message: '长度在1~100位之间', trigger: 'blur'},
+        ],
+      }
     }
   },
   created() {
@@ -174,37 +160,28 @@ export default {
   },
   methods: {
     details(row) {
-      this.detail = row
+      this.detail = row;
       this.vis = true
     },
     reset() {
-      request.get("/events", {
-        params: {
-          pageNum: this.currentPage,
-          pageSize: this.pageSize,
-          title: '',
-          author: '',
-          // createTime: ''
-        }
-      }).then(res => {
-        // console.log(res);
-        this.tableData = res.data.records;
-        this.total = res.data.total;
-        this.title = '';
-        this.author = '';
-      })
+      this.title = '';
+      this.author = '';
+      this.timeHorizon = [];
+      this.load();
     },
     load() {
+      let time = this.timeHorizon;
+      // console.log(time);
       request.get("/events", {
         params: {
           pageNum: this.currentPage,
           pageSize: this.pageSize,
           title: this.title,
           author: this.author,
-          // createTime: this.createTime
+          startTime: time[0],
+          endTime: time[1]
         }
       }).then(res => {
-        // console.log(res);
         this.tableData = res.data.records;
         this.total = res.data.total;
       })
@@ -219,8 +196,8 @@ export default {
           //关联add弹窗里面的div，new一个editor对象
           editor = new E('#div1');
           //本地图片上传设置注意后端设置返回json格式
-          editor.config.uploadImgServer = 'http://localhost:9090/files/editor/upload'
-          editor.config.uploadFileName = "file"
+          editor.config.uploadImgServer = 'http://' + serverIp + ':9090/files/editor/upload';
+          editor.config.uploadFileName = "file";
           editor.create()
         }
         editor.txt.html("")
@@ -228,41 +205,38 @@ export default {
 
     },
     save() {
-      this.form.content = editor.txt.html(); //获取编辑器里面的值。然后赋予到实体form对象当中
+      this.$refs.pass.validate((valid) => {
+        if (valid) {
+          this.loading = true
+          setTimeout(() => {
+            this.loading = false
+          }, 1000)
 
-      if (this.form.id) {//若果id存在，更新
-        let userStr = sessionStorage.getItem("user") || "{}"
-        let user = JSON.parse(userStr)
-        // console.log("原作者："+this.form.author+",新作者："+user.name)
-        if (this.form.author === user.name) {
-          request.put("/events", this.form).then(res => {
+          this.form.content = editor.txt.html(); //获取编辑器里面的值。然后赋予到实体form对象当中
+          // console.log(typeof this.form.content)
+          if (this.form.content === "") {
+            this.$message.error("正文不能为空")
+            return false;
+          }
+          this.form.author = this.user.name;
+          request.post("/events/", this.form).then(res => {
             // console.log(res);
             if (res.code === '0') {
               this.$message.success("编辑成功")
+              this.load();//刷新表格数据
+              this.dialogVisible = false
             } else {
               this.$message.error(res.msg)
             }
-            this.load();//刷新表格数据
-            this.dialogVisible = false
           });
-        } else {
-          this.$message.error("不是作者本人，编辑无效！")
         }
+      })
 
-      } else {//如果id不存在，新增
 
-        this.form.author = user.name
-        request.post("/events", this.form).then(res => {
-          // console.log(res);
-          if (res.code === '0') {
-            this.$message.success("发表成功")
-          } else {
-            this.$message.error(res.msg)
-          }
-          this.load();//刷新表格数据
-          this.dialogVisible = false
-        });
-      }
+      // else {
+      //     this.$message.error("不是作者本人，编辑无效！")
+      // }
+
     },
     handleEdit(row) {
       this.form = JSON.parse(JSON.stringify(row));
@@ -271,34 +245,61 @@ export default {
       this.$nextTick(() => {
         if (!editor) {
           editor = new E('#div1');
-          editor.config.uploadImgServer = 'http://localhost:9090/files/editor/upload'
-          editor.config.uploadFileName = "file"
+          editor.config.uploadImgServer = 'http://' + serverIp + ':9090/files/editor/upload';
+          editor.config.uploadFileName = "file";
           editor.create()
         }
         editor.txt.html(row.content)
       })
-    },
+    }
+    ,
     handleDelete(row) {
+      // console.log(this.user)
+      if (row.author === this.user.name || this.user.role === "ROLE_ADMIN") {
+        this.id = row.id;
+        request.delete("/events/" + this.id).then(res => {
+          if (res.code === '0') {
+            this.$message.success("删除成功")
+          } else {
+            this.$message.error(res.msg)
+          }
+          this.load();//刷新表格数据
+        })
+      } else {
+        this.$message.error("不是你创建的不能删哦")
+      }
 
-      this.id = row.id;
-      // console.log(this.id);
-      request.delete("/events/" + this.id).then(res => {
-        if (res.code === '0') {
-          this.$message.success("删除成功")
-        } else {
-          this.$message.error(res.msg)
-        }
-        this.load();//刷新表格数据
-      })
-    },
+    }
+    ,
     handleSizeChange() {
       //改变当前每页个数的触发
       this.load()
-    },
+    }
+    ,
     handleCurrentChange() {
       //改变当前页数的触发
       this.load()
-    },
+    }
+    ,
+    deleteBatch() {
+      if (!this.ids.length) {
+        this.$message.warning("请选择要删除的事件");
+        return
+      }
+      request.post("/events/deleteBatch", this.ids).then(res => {
+        if (res.code === '0') {
+          this.$message.success("批量删除成功")
+          this.load()
+        } else {
+          this.$message.error(res.msg)
+        }
+      })
+    }
+    ,
+    handleSelectionChange(val) {
+      this.ids = val.map(v => v.id)
+    }
+    ,
   }
 }
 </script>
